+++ src/main/java/veterinaria/service/MascotaService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface MascotaService {
     public Mascota creatMascota(Mascota mascota);
-    public Mascota findMascota(Integer id);
+    public Mascota findMascota(Long id);
     public List<Mascota> findAllMascotas();
-    public Mascota updateMascota(Mascota mascota ,Integer id);
-    public Mascota deleteMascota(Integer id);
+    public Mascota updateMascota(Mascota mascota ,Long id);
+    public Mascota deleteMascota(Long id);
 }
